requirements.txt
flask
flasgger
tensorflow
pillow
numpy
keras
gunicorn